Add logging configuration function to start.py
This commit will add start.py `configure_logging()`. The function accepts
a `Path()` to a logging configuration file, which can be either .py,
.yml, .yaml, .conf, or .ini.

For .py files, the function will load the Python file as a module, load
the `LOGGING_CONFIG` dictionary object, and pass it to
`logging.config.dictConfig()`.

YAML files are opened with a context manager. PyYAML parses the YAML, and
then passes the parsed YAML dictionary to `logging.config.dictConfig()`.

The old-style .conf, or .ini. files are loaded with
`logging.config.fileConfig()`.

The type: ignore mypy comment prevents the following error messages,
related to `spec.loader.exec_module()`:

- Item "_Loader" of "Optional[_Loader]" has no attribute "exec_module"
  [union-attr] mypy(error)
- Item "None" of "Optional[_Loader]" has no attribute "exec_module"
  [union-attr] mypy(error)

https://mypy.readthedocs.io/en/latest/error_code_list.html