Retain package directory structure in Docker image
Prior to this commit, files from the inboard package were broken up and
added to the root directory of the Docker image, to allow users to
overwrite the /app directory without affecting the running Docker image.
This led to some limitations:

- Complicated `COPY` commands in the Dockerfile
- Inability to use package imports in package modules: gunicorn_conf.py
  needed `from start` instead of `from inboard.start`. This limits the
  usefulness of inboard as a Python package.

This commit will restructure the inboard repo to retain the inboard
directory in the Docker image, at /app/inboard, simplifying many of the
file and module paths as a result. This will also simplify operations
for downstream applications, because imports stay the same for Python
packages with or without Docker.

There's a risk that users who were overwriting /app when using
tiangolo/uvicorn-gunicorn will accidentally overwrite the /app directory
here, so the README will be updated to clarify the preferred setup.

Changes:

- Update README to encourage users to copy downstream application files
  into /app/package instead of /app
- Update `set_conf_path()` to return `str` instead of `Path`, allowing
  `configure_logging()` to read the `LOGGING_CONF` environment variable
  as a default argument
- Update file and module paths for new directory structure
- Add exception handling to inboard/`__init__.py` version calculation:
  Although the inboard directory is now copied into the Docker image,
  the package is still not installed, to allow users to overwrite the
  pyproject.toml and poetry.lock with their own. The version calculation
  in inboard/`__init__.py` depends on `__package__`, which is not
  present unless inboard is fully installed as a package.