Move latest tag to final workflow step
br3ndonland/inboard@21c3bcd

This commit provides a minor update to the order in which Docker images
are built, tagged, and pushed by GitHub Actions. The `latest` tag
will be moved to a separate workflow step, to ensure it always shows up
as the latest version in GitHub Container Registry.