Improve Gunicorn performance auto-tuning
tiangolo/uvicorn-gunicorn-docker#5
tiangolo/uvicorn-gunicorn-starlette-docker#4
tiangolo/uvicorn-gunicorn-fastapi-docker#6

The "auto-tuning" advertised in tiangolo/uvicorn-gunicorn-docker is
basically a few lines of the `gunicorn_conf.py` that determine the
number of Gunicorn workers to run. It would be helpful to write some
unit test cases for this feature, but without being in a separate unit,
it is difficult to unit test in isolation.

This commit will refactor the performance auto-tuning into a function,
`gunicorn_conf.calculate_workers`, and will add unit test cases to
verify the resulting number of worker processes.