Fix web concurrency Gunicorn worker calculation
br3ndonland/inboard#11

The README explains, "If either `MAX_WORKERS` or `WEB_CONCURRENCY` are
set to 1, the total number of workers will be 1, overriding the default
minimum of 2." This was true for `MAX_WORKERS`, but `WEB_CONCURRENCY`
was still set to minimum 2.

This commit will allow `WEB_CONCURRENCY` (number of Gunicorn workers) to
have a minimum of 1.