Parse pyproject.toml for testing
So far, testing has been limited to the Python modules. However, the
pyproject.toml is an important source of metadata for the project. This
commit will use toml (pytest dependency) to load pyproject.toml, and
Pydantic (FastAPI dependency) will be used for data validation. Unit
tests will verify the presence of select fields in the pyproject.toml.

It is also possible to use `importlib.metadata` to read other package
metadata, as explained in python-poetry/poetry#1036. The metadata
originate in pyproject.toml, and becomes available after Poetry
installs the root project package. The `importlib.metadata` approach
works in a virtualenv, in which the Poetry root project is installed,
but is error prone in Docker, because the root project is usually not
installed. Parsing the TOML is a more durable alternative, because it
works even if the root project is not installed.

The `importlib.metadata` version calculation in `inboard/__init__.py`
was not necessary, so it will be removed.