Remove Docker SDK
https://docker-py.readthedocs.io/en/stable/

Eventually, I hope to run some integration tests, in which I check the
logs of the running Docker containers. I have tried working with the
Docker SDK for Python, but I haven't been happy with the slow speed and
inconsistent performance. This commit will remove the Docker SDK.

Requests is still required for the FastAPI/Starlette TestClient.