Use Python version matrix in Docker job
br3ndonland/inboard#19

Commit dc8da18 updated the builds.yml GitHub Actions workflow to build
Docker images for multiple Python versions. The initial strategy was to
output the Python version used to the Docker job, so that the Docker job
depends on the success of the Python job with the same version.

This may not be possible. If it is, it's unclear how to set a downstream
job to depend on one specific matrix job (for example, 3.8 in the python
job, instead of just the python job in general).

https://docs.github.com/en/free-pro-team@latest/actions/reference/context-and-expression-syntax-for-github-actions#needs-context