Enable Starlette 0.14
As far as I can tell, it's not possible to install both FastAPI 0.62 and
Starlette 0.14 in the same Poetry project. FastAPI 0.62.0 has Starlette
pinned at 0.13.6, and attempts to add Starlette 0.14 raise a Poetry
`SolverProblemError` (on Poetry 1.1.4):

```text
❯ poetry add starlette@'^0.14'

Updating dependencies
Resolving dependencies... (0.1s)

  SolverProblemError

  Because fastapi (0.62.0) depends on starlette (0.13.6)
  and no versions of fastapi match >0.62,<0.63, fastapi (>=0.62,<0.63)
  requires starlette (0.13.6).
  So, because inboard depends on both fastapi (^0.62) and starlette
  (^0.14), version solving failed.

  at ~/.poetry/lib/poetry/puzzle/solver.py:241 in _solve
      237│             packages = result.packages
      238│         except OverrideNeeded as e:
      239│             return self.solve_in_compatibility_mode(e.overrides, use_latest=use_latest)
      240│         except SolveFailure as e:
    → 241│             raise SolverProblemError(e)
      242│
      243│         results = dict(
      244│             depth_first_search(
      245│                 PackageNode(self._package, packages), aggregate_package_nodes
```

I would be open to having FastAPI 0.62 and Starlette 0.14 be separate
sets of dependencies, but even if only one set of extras is installed
(`poetry install -E starlette`), Poetry appears to resolve all the
dependencies together, so the environment will still end up with
Starlette 0.13.6.

The easiest solution is to just install the Starlette version pinned by
FastAPI (0.13.6), but add a Poetry union operator (`||`) so upgrading
to Starlette 0.14 is at least possible in the future. The order of the
union doesn't seem to matter.

In Docker, FastAPI can be removed from the Poetry project prior to
installing Starlette, so that Starlette 0.14 can be installed.