Implement Uvicorn reload_dirs
https://github.com/samuelcolvin/watchgod
https://github.com/encode/uvicorn/releases/tag/0.11.4
https://www.uvicorn.org/settings/

watchgod is a filesystem watcher that was added in Uvicorn 0.11.4. The
directories to watch can be configured from the command-line with the
`--reload-dir` argument. Multiple `--reload-dir` can be used, with one
directory specified each time.

The Uvicorn docs say, "If you're running using programmatically, using
`uvicorn.run(...)`, then use equivalent keyword arguments," but this is
not quite accurate for `--reload-dir`. The programmatic equivalent is
`reload_dirs`, which accepts `List[str]`.

To implement `reload_dirs` in a customizable fashion, the `RELOAD_DIRS`
environment variable will be added, which should be a comma-separated
string. The string will then be parsed, and the directory names passed
to Uvicorn. Directory names are relative to the project root. The string
is parsed with `"".split(sep=",")`, and any leading whitespace removed
with `"".lstrip()`. This approach is preferred over `"".split(sep=", ")`
in case users forget to add a space between the directories.