Fix HTML syntax in README
br3ndonland/inboard@c217e9b

The centered elements in the README are showing up as expected on GitHub
but not on PyPI. This commit will add some minor corrections to the HTML
syntax. If it's still not showing up centered on PyPI after this, it's
not a big deal.