Correct GitHub Actions Docker push step
br3ndonland/inboard#19
https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions

inboard provides a variety of Docker tags. The least specific tags are:

- ghcr.io/br3ndonland/inboard
- ghcr.io/br3ndonland/inboard:base
- ghcr.io/br3ndonland/inboard:fastapi
- ghcr.io/br3ndonland/inboard:starlette

Images with these tags should only be pushed if they are built with the
latest Python version and the Git ref is either a tag, develop, or main.
Docker images built with Python 3.8 were still being pushed with these
tags, because the syntax evaluated to true if only one of the conditions
was true.

This commit will add a logical grouping to the if statement to correct
the Docker push behavior.