Test Gunicorn settings
br3ndonland/inboard#11
https://docs.gunicorn.org/en/latest/settings.html#print-config
https://docs.pytest.org/en/latest/how-to/capture-stdout-stderr.html

The Gunicorn configuration file was technically covered by unit tests,
but the individual settings were not tested. This commit will add tests
of the Gunicorn settings. The tests use Gunicorn's `--print-config`
option to load and output the configuration. The pytest fixture `capfd`
is used to read the Gunicorn output.