Split up logging configuration method
br3ndonland/inboard#3

The `configure_logging()` method did multiple things: find and import a
logging configuration module, load the logging configuration dictionary,
and apply the dictionary configuration to the Python logger.

This commit will refactor `configure_logging()` into two methods.

1. `find_and_load_logging_conf()` will return a dictionary configuration
2. `configure_logging()` will apply the dictionary configuration

The return value of `configure_logging()` will remain the same, so there
will not be any changes to the programming API.

Also, now that the `configure_logging()` method is in the same module as
the logging configuration dictionary, that dictionary will be used by
default, instead of finding the module separately.