Test logger output after configuring logging
br3ndonland/inboard#3
https://docs.pytest.org/en/latest/how-to/capture-stdout-stderr.html
https://docs.pytest.org/en/latest/how-to/logging.html

In addition to testing the methods that load the logging configuration,
it can also be useful to test that logger messages have the expected
format. This commit will add tests that output logger messages in
various formats and assert that each output matches the expected format.

As in the tests of the Gunicorn server output, these tests will use the
pytest `capfd` fixture. It is also possible to use the `caplog` fixture,
but I have not had as much success with it.