Add __all__ to __init__.py
https://docs.python.org/3/reference/simple_stmts.html#the-import-statement

In Python, `__all__` is a sequence (typically a list or tuple) that
defines the public names of a module or package. When other applications
install the package, the objects corresponding to the public names are
exposed for direct import and use.

This commit will add `__all__`, defining the public API of the package.
To start, the public API will consist of a few useful utilities that are
commonly imported in other Python applications, such as the logging
configuration logic and HTTP Basic auth implementations.