Update mypy configuration file
https://mypy.readthedocs.io/en/stable/command_line.html
https://mypy.readthedocs.io/en/stable/config_file.html
https://mypy.readthedocs.io/en/stable/running_mypy.html#mapping-paths-to-modules

This commit will serve as a follow-up commit to 96132a6, by adding the
`explicit_package_bases` setting, which is a companion to the PEP 420
`namespace_packages` setting.

Enabling both PEP 420 namespace packages and explicit package bases
allows relative imports when the top-level package lacks a `__init__.py`
file, such as for imports within the tests directories.