Only use Gunicorn max_workers setting as a limit
br3ndonland/inboard#31
br3ndonland/inboard@315c1c1

Commit 315c1c1 improved the calculation of the number of Gunicorn worker
processes started by the inboard web server.

The `gunicorn_conf.calculate_workers()` method was update to return
`use_least or use_max or use_total or use_default`. However, this meant
that `use_max`, which is set by the `MAX_WORKERS` environment variable,
could potentially be used as the total number of workers, when it should
only be used as a maximum, not as a total.

This commit will update `gunicorn_conf.calculate_workers()` and related
tests so that `MAX_WORKERS` is only used to set a limit on the total,
not to set the total itself. The `gunicorn_conf.calculate_workers()`
method will now return `use_least or use_total or use_default`.