Update type annotations for Python 3.10
This commit will update type annotation syntax for Python 3.10. The
project currently also supports Python 3.8 and 3.9, so the annotations
are imported with `from __future__ import annotations`.

The Python 3.10 union operator (the pipe, like `str | None`) will not be
used on pydantic models. If running Python 3.9 or below, pydantic is not
compatible with the union operator, even if annotations are imported
with `from __future__ import annotations`.

https://peps.python.org/pep-0604/
https://docs.python.org/3/whatsnew/3.10.html
https://github.com/samuelcolvin/pydantic/issues/2597#issuecomment-1086194186
https://github.com/samuelcolvin/pydantic/pull/2609#issuecomment-1084341812
https://github.com/samuelcolvin/pydantic/issues/3300#issuecomment-1034007897