Upgrade to Uvicorn 0.14 and Click 8.0
https://github.com/encode/uvicorn/releases/tag/0.14.0
https://click.palletsprojects.com/en/8.0.x/changes/

Uvicorn now allows Click 8 to be used (encode/uvicorn#1042), which is
notable because of a security fix added in Click 8. This fix patches an
insecure temporary file vulnerability, and was reported by CodeQL scans
in the inboard repo. The Click maintainers did not clearly report this
vulnerability in the release notes. The release notes entry merely says,
"Use `mkstemp()` instead of `mktemp()` in pager implementation." The fix
provided in pallets/click#1754 does not explain the implications of the
change. The issue report in pallets/click#1752 was criticized for not
respecting https://github.com/pallets/click/security/policy, but Click
did not provide adequate information otherwise.