Use standard input for GitHub Actions Docker login
br3ndonland/inboard#5
https://docs.docker.com/engine/reference/commandline/login/
https://docs.github.com/en/actions/reference/encrypted-secrets

This commit will update the Docker login syntax used in GitHub Actions
to the `--password-stdin` syntax, which ensures that the password is not
directly entered on the command-line.

GitHub Actions obfuscates credentials stored in GitHub Secrets, so there
was never any issue with passwords being exposed. This is simply a
best-practice adjustment that provides an extra layer of security.