Drop `toml` dependency
https://toml.io/en/
uiri/toml#267
https://github.com/uiri/toml/pull/279#issuecomment-633012586

While the TOML spec is alive and well, the `toml` package that was used
in this project is dead (uiri/toml#267). There are several alternatives.
`rtoml` is the fastest, but does not provide "round trip" guarantees
(meaning that it can't load, then dump, and get an identical result).
However, for simple loading and parsing it's fine. `tomli` is another
alternative, but it is read-only (writing requires a separate `tomli-w`
package), and requires files to be opened in binary mode to parse TOML.
`tomlkit` is used by Poetry (from the Poetry author), but is currently
70x slower than `rtoml`.

It's also important to note that the TOML parsing and settings loading
is somewhat tangential to the focus of the inboard project. A separate
project, https://github.com/br3ndonland/fastenv, is more focused on
settings management, and may feature TOML support in the future.

This commit will remove `toml` from the `fastapi` Poetry extras group.
Note that `toml` is still a sub-dependency of some `dev-dependencies`,
including pre-commit and pytest.