Use consistent substring match for Poetry version
br3ndonland/inboard#36
br3ndonland/inboard#44

PR #44 introduced version tests for Poetry. The GitHub Actions workflow
builds.yml was matching the version number as a substring anywhere in
the output of `poetry -V`, but the hooks.yml and tests.yml workflows
were matching the version number at the end of the `poetry -V` output.
This would have become an issue if the format of the output was changed,
and sure enough, Poetry 1.2.0 changes the format to have parentheses
around the version number.

This commit will update the hooks.yml and tests.yml workflows to match
the version number as a substring anywhere in the output of `poetry -V`.