Check pre-start script exit code
br3ndonland/inboard#2

inboard optionally runs a pre-start script before starting the server.
The path to a pre-start script can be specified with the environment
variable `PRE_START_PATH`. If the environment variable is set to a
nonzero value, inboard will run the script at the provided path, using
the `subprocess` standard library package.

Previously, if the script exited with an error, inboard would continue
starting the server. However, it may be preferable to stop the server
if the pre-start script fails. This commit will update the subprocess
call to include `check=True`. If the pre-start script exits with an
error, inboard will not start the server.

The behavior of successful pre-start script runs will not change.
However, failed pre-start script runs will now exit with error codes,
and prevent the server from starting.