Add complete logging override example to docs
inboard propagates FastAPI, Gunicorn, and Uvicorn log streams to the
root logger, so that all the log streams can be handled by the same
logging config in a simple way.

Users can modify inboard's logging behavior, or override it completely.
To override inboard's entire logging config, set up a `LOGGING_CONFIG`
dictionary, and pass inboard the path to the module with the dictionary.

This commit will add a complete `LOGGING_CONFIG` example to the docs,
with some common customizations that may be of interest to users.