Add configurable logging filter
Filters identify log messages to filter out, so that the logger does not
log messages containing any of the filters. If any matches are present
in a log message, the logger will not output the message.

The environment variable `LOG_FILTERS` can be used to specify filters as
a comma-separated string, like `LOG_FILTERS="/health, /heartbeat"`. To
then add the filters to a class instance, the `LogFilter.set_filters()`
method can make the set of filters from the environment variable value.

One of the primary use cases for log message filters is health checks.
When applications with APIs are deployed, it is common to perform
"health checks" on them. Health checks are usually performed by making
HTTP requests to a designated API endpoint. These checks are made at
frequent intervals, and so they can fill up the access logs with large
numbers of unnecessary log records. To avoid logging health checks,
add those endpoints to the `LOG_FILTERS` environment variable.