Update logging filter docs to use Docker image
br3ndonland/inboard#49
https://inboard.bws.bio/logging

The docs on filtering log messages were initially written with examples
that used the VSCode debugger. This commit will update the docs to use
one of the inboard Docker images instead. This change makes the docs
more relevant to users who may be using the inboard Docker images, but
not working with the inboard source code directly.