Add `BUILDKIT_INLINE_CACHE` Docker build argument
The `BUILDKIT_INLINE_CACHE` build argument tells Docker to write cache
metadata into the image during `docker build`. The image can then be
used as a cache source for subsequent builds.

The build argument will be provided with the `docker build` command.
The build argument could be provided directly within the Dockerfile,
but BuildKit does not offer any guidance on where it goes. It could
either be supplied before the first `FROM`, or after the first `FROM`,
and could work differently in each case.

https://docs.docker.com/engine/reference/commandline/build/#specifying-external-cache-sources