Update to FastAPI 0.87 and Starlette 0.21
This commit will update/upgrade to FastAPI 0.87 and Starlette 0.21.

BREAKING CHANGE: Starlette 0.21 updated the test client to run on HTTPX
instead of Requests. Please verify compatibility with your test suite
after updating. Also note that, while Requests is no longer a dependency
of Starlette, it recently became a dependency of mkdocs-material, so it
will still be included when installing mkdocs-material. mkdocs-material
(aka "Material for MkDocs") is used to build the documentation. They
recently achieved their $8000 "Scotch Bonnet" funding goal, and released
a collection of previously insiders-only features to the public. One of
these features is a new built-in social plugin that uses Requests.

https://github.com/tiangolo/fastapi/releases/tag/0.87.0
https://github.com/encode/starlette/releases/tag/0.21.0
https://www.python-httpx.org/compatibility/
https://squidfunk.github.io/mkdocs-material/insiders/
https://squidfunk.github.io/mkdocs-material/setup/setting-up-social-cards/