Enable `pymdownx.magiclink`
This commit will enable the `MagicLink` Python Markdown extension.

This extension will "magic link" (autolink) GitHub references, meaning
that it will add hyperlinks to the text on the page.

This extension is not officially supported by Material for MkDocs, but
it appears to work well, and it helps add links to the changelog.

https://facelessuser.github.io/pymdown-extensions/extensions/magiclink/
https://squidfunk.github.io/mkdocs-material/setup/extensions/python-markdown-extensions/
https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/autolinked-references-and-urls