Update to mypy 1
This commit will update to mypy 1. In the past, releases were managed by
Dropbox (python/mypy#12210), and were done at infrequent intervals. They
plan to release more regularly in the future.

Mypy does not keep a changelog, but did provide a 1.0 release blog post:
https://mypy-lang.blogspot.com/2023/02/mypy-10-released.html

A changelog has been requested (python/mypy#2859, python/mypy#13279),
but has been dismissed because "you can view the git history easily."

The new release versioning scheme is in the format `major.minor.patch`,
but is not SemVer-compliant, as explained in their blog post and wiki.
https://github.com/python/mypy/wiki/Release-Process

The update to mypy 1 was postponed until bug fixes were available in the
1.0.1 patch release (python/mypy#13685).