Update to recommended Starlette middleware syntax
Starlette is no longer documenting (but not yet deprecating) the
`app.add_middleware` syntax in favor of passing a middleware list
as a kwarg to `Starlette()` (`Starlette(middleware=middleware)`). The
recommended syntax allows for more predictable ordering of middleware
lists, so that they are added to the Starlette app and initialized in
expected order.

Use of the `app.add_middleware` syntax in the inboard example Starlette
and FastAPI apps will be replaced with the recommended approach.