Relate inboard, FastAPI, and Uvicorn versions
Each minor version release of FastAPI or Uvicorn gets a corresponding
minor version release of inboard.

FastAPI depends on Starlette, so any updates to the Starlette version
required by FastAPI should be included with updates to FastAPI.

The CHANGELOG.md entries explain this practice by including notes like
"This is a minor release to align with FastAPI and Starlette versioning"
each time a release is created.