Remove Poetry 1.1 from Docker images
As described in the changelog entry for
[inboard 0.38](https://inboard.bws.bio/changelog#0380-2023-02-26) and
the [inboard docs](https://inboard.bws.bio/docker#docker-and-poetry),
inboard switched its dependency management and packaging from Poetry 1.1
to Hatch. Poetry 1.1 was retained in the inboard Docker images for
backwards compatibility, but Poetry 1.1 is unmaintained and so it must
eventually be removed.

This commit will remove Poetry 1.1 from the inboard Docker images.
This is a **BREAKING CHANGE**.

How to deal with this breaking change:

- If you are not using Poetry there are no changes needed.
- If you are still using Poetry add `RUN pipx install poetry`
  to your Dockerfile.