Update to FastAPI 0.101
This commit will update/upgrade to
[FastAPI 0.101](https://fastapi.tiangolo.com/release-notes/).

FastAPI 0.101 adds support for
[Pydantic computed fields](https://docs.pydantic.dev/latest/usage/computed_fields/)
and alters responses to use
[Pydantic serialization](https://docs.pydantic.dev/latest/usage/serialization/).
The change to Pydantic serialization in FastAPI 0.101 separates input
and output schemas, altering responses and OpenAPI schema generation.
Due to the lack of backwards compatibility and the associated issues
(https://github.com/tiangolo/fastapi/pull/10011#issuecomment-1676624793,
https://github.com/tiangolo/fastapi/discussions/10019,
https://github.com/tiangolo/fastapi/discussions/10041),
the update to FastAPI 0.101 may be a **BREAKING CHANGE** for some users.

How to deal with this breaking change:

- If this change affects your project adversely, skip FastAPI 0.101 and
  update to FastAPI 0.102 (upcoming in the next inboard release).
- Set `separate_input_output_schemas=False` on the FastAPI app instance
  (`app = FastAPI(separate_input_output_schemas=False)`) as shown in the
  [FastAPI docs](https://fastapi.tiangolo.com/how-to/separate-openapi-schemas/).