Update to FastAPI 0.102
This commit will update/upgrade to
[FastAPI 0.102](https://fastapi.tiangolo.com/release-notes/).

FastAPI 0.101 added support for
[Pydantic computed fields](https://docs.pydantic.dev/latest/usage/computed_fields/)
and altered responses to use
[Pydantic serialization](https://docs.pydantic.dev/latest/usage/serialization/).
The change to Pydantic serialization in FastAPI 0.101 separated input
and output schemas, altering responses and OpenAPI schema generation.
Due to the lack of backwards compatibility and the associated issues
(https://github.com/tiangolo/fastapi/pull/10011#issuecomment-1676624793,
https://github.com/tiangolo/fastapi/discussions/10019,
https://github.com/tiangolo/fastapi/discussions/10041),
the update to FastAPI 0.101 may be a **BREAKING CHANGE** for some users.

FastAPI 0.102 builds on the serialization changes introduced in 0.101 by
adding docs explaining the serialization behavior and by adding a new
optional keyword argument for disabling the new behavior.

If the Pydantic serialization behavior affects your project adversely,
disable it with `separate_input_output_schemas=False` on the FastAPI app
(`app = FastAPI(separate_input_output_schemas=False)`) as shown in the
[FastAPI docs](https://fastapi.tiangolo.com/how-to/separate-openapi-schemas/).