Update to FastAPI 0.106
This commit will update/upgrade to
[FastAPI 0.106](https://fastapi.tiangolo.com/release-notes/).

In FastAPI 0.74 (released in inboard 0.18.0 - 2022-03-05), the internal
`AsyncExitStack` was updated so that dependencies with `yield` could
catch exceptions like `HTTPException`.

FastAPI 0.106 builds on the 0.74 updates by introducing the ability to
raise exceptions after `yield`. This update includes a BREAKING CHANGE
because objects from dependencies with `yield` can no longer be used in
background tasks. The recommendation in the
[FastAPI docs](https://fastapi.tiangolo.com/tutorial/dependencies/dependencies-with-yield/)
for updating background tasks is:

> If you used to rely on this behavior, now you should create the
> resources for background tasks inside the background task itself, and
> use internally only data that doesn't depend on the resources of
> dependencies with `yield`.
>
> For example, instead of using the same database session, you would
> create a new database session inside of the background task, and you
> would obtain the objects from the database using this new session. And
> then instead of passing the object from the database as a parameter to
> the background task function, you would pass the ID of that object and
> then obtain the object again inside the background task function.