Update to Uvicorn 0.25.0
This commit will update/upgrade to
[Uvicorn 0.25.0](https://github.com/encode/uvicorn/releases).

Uvicorn 0.25.0 adds support for the WebSocket Denial Response ASGI
extension. This is used in certain cases in which a WebSocket app needs
to reject a connection and return a custom response.

Uvicorn 0.25.0 also includes some corrections to the type annotations on
`uvicorn.run()`. `inboard.types.UvicornOptions` already included correct
type annotations that match these corrections, so no changes are needed.