Avoid `metadata-generation-failed` in Dockerfiles
The example Dockerfiles in the docs raise a `metadata-generation-failed`
error during the `hatch env` commands. The error is occurring because
the `pyproject.toml` has a dynamic version that is stored in
`package_name/__init__.py`, but the source code in `package_name` hasn't
been copied into the Docker image yet.

The reason the Dockerfile was set up to `COPY` the source code after
installing the environment is to leverage the Docker build cache.
To continue leveraging the Docker build cache while moving the `COPY`
command earlier, `COPY --link` syntax can be used. This syntax was
introduced in Dockerfile 1.4.0.

This commit will update the docs to move the `COPY --link` step earlier
to avoid the error.

https://docs.docker.com/build/cache/
https://github.com/br3ndonland/inboard/discussions/98
https://github.com/moby/buildkit/releases/tag/dockerfile%2F1.4.0