Update to FastAPI 0.110 and Starlette 0.36
This commit will update/upgrade to
[FastAPI 0.110](https://fastapi.tiangolo.com/release-notes/)
and
[Starlette 0.36](https://www.starlette.io/release-notes/).
FastAPI 0.110 makes a change to dependencies with `yield` and `except`.
Dependencies must now raise exceptions after `except`. This change is
intended to address memory leak issues and may be a breaking change in
some projects if dependencies with `yield` and `except` used `pass`
instead of `raise`. See the
[FastAPI docs](https://fastapi.tiangolo.com/tutorial/dependencies/dependencies-with-yield/)
for further info.

Changes to Starlette between 0.35 and 0.36 include exception handling
updates and AnyIO compatibility updates. Note that FastAPI updated the
Starlette minor version from 0.35 to 0.36 in the 0.109.2 patch release.