Publish to PyPI with OIDC trusted publisher
This commit will update Python package publishing to the newest format
recommended by PyPI. This project previously published packages with the
`hatch publish` command and a project-scoped PyPI API token (token only
valid for this project) stored in GitHub Secrets. The project will now
publish packages using a PyPI OIDC (OpenID Connect) trusted publisher
with the pypa/gh-action-pypi-publish action. This is the method that
Hatch uses (Hatch does not "dogfood" its own `hatch publish` feature).

The advantage to OIDC is that authentication is performed with temporary
API tokens (only valid for 15 minutes) instead of persistent tokens that
must be manually generated on PyPI and pasted into GitHub Secrets. The
disadvantage is that authentication is more complicated.

To use PyPI OIDC, a trusted publisher was set up for the PyPI project
as shown in the [PyPI docs](https://docs.pypi.org/trusted-publishers/).
Next, a dedicated GitHub Actions deployment environment was created for
PyPI, with protection rules that only allow use of the environment with
Git tags. The environment protection rules combine with tag protection
rules in the existing GitHub rulesets to ensure PyPI packages can only
be published if a maintainer triggers a workflow run with a Git tag ref.

The GitHub Actions workflow will now be updated to use the deployment
environment. Deployment environments must be selected at the job level
before the job begins, so a setup job will be added that selects the
appropriate deployment environment and passes it to the PyPI job.

Each use of a deployment environment creates a deployment that can be
either active or inactive. The GitHub REST API docs for deployments say,
"When you set the state of a deployment to `success`, then all prior
non-transient, non-production environment deployments in the same
repository with the same environment name will become `inactive`. To
avoid this, you can set `auto_inactive` to `false` when creating the
deployment status."
https://docs.github.com/en/rest/deployments/deployments?apiVersion=2022-11-28

GitHub Actions auto-inactivates deployments. When a deployment
environment is used from within a GitHub Actions workflow job with the
`environment:` configuration syntax, GitHub Actions marks previous
deployments to that environment as `inactive`. This behavior is
undocumented and there doesn't appear to be a way to configure it, but
there are some possible workarounds/hacks. One workaround could be using
the GitHub deployment statuses API to reactivate deployments of interest.
https://docs.github.com/en/rest/deployments/statuses?apiVersion=2022-11-28
Another workaround for keeping deployments active could be through
deployment URLs. A GitHub community discussion comment hints, "If the
new Deployment specifies an environment URL in the workflow definition,
only previous Deployments with the exact same environment URL will be
auto-inactivated." This implies that deployments could stay active if
each deployment got its own unique URL. This approach will be used.
https://github.com/orgs/community/discussions/67982#discussioncomment-7086962

To publish the Python package to PyPI, `hatch build` will output package
build files to the `dist/` directory, then pypa/gh-action-pypi-publish
will authenticate and upload the files. pypa/gh-action-pypi-publish
provides exact version tags like pypa/gh-action-pypi-publish@v1.8.14 and
branches for major and minor version numbers like
pypa/gh-action-pypi-publish@release/v1.8.

https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-pypi
https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-rulesets/about-rulesets
https://docs.pypi.org/trusted-publishers/
https://github.com/pypa/gh-action-pypi-publish